+++ index.js
@@ -1,9 +1,9 @@
 /**
- * @Author: abbeymart | Abi Akindele | @Created: 2019-01-12 | @Updated: 2019-06-02
+ * @Author: abbeymart | Abi Akindele | @Created: 2019-01-12 | @Updated: 2019-06-05
  * @Company: mConnect.biz | @License: MIT
  * @Description: @mconnect/validate-crud | CRUD activities-params/args validation
  */
 
-const validateCrud = require('./src/validateCRUD');
+const {validateCrud, ValidateCrud} = require('./src/validateCRUD');
 
-module.exports = validateCrud;
+module.exports = {validateCrud, ValidateCrud};