+++ HISTORY.md
@@ -1,5 +1,11 @@
 # Historical logs
 
+## Version 0.3.0
+- Released June 05, 2019
+- fixed/updated options' settings
+- added proto-type constructor/instance function feature
+
+
 ## Version 0.2.1
 - Released June 03, 2019
 - resolved mcMessages naming conflict / initialization issue